--- apps/fewd/components/resources-per-week-table-row.tsx
@@ -1,28 +1,0 @@
-import Link from 'next/link';
-import React from 'react';
-
-const ResourcesPerWeekTableRow: React.FC<Week> = ({
-  weekNumber,
-  lesson1,
-  lesson2,
-  homework,
-}) => {
-  return (
-    <tr>
-      <td>
-        <Link href={`week/${weekNumber}`}>{weekNumber}</Link>
-      </td>
-      <td>
-        <Link href={`/lessons/${lesson1.link}`}>{lesson1.content}</Link>
-      </td>
-      <td>
-        <Link href={`/lessons/${lesson2.link}`}>{lesson2.content}</Link>
-      </td>
-      <td>
-        {homework ? <Link href={homework.link}>{homework.content}</Link> : ''}
-      </td>
-    </tr>
-  );
-};
-
-export default ResourcesPerWeekTableRow;
